#465 migrate helper funcs from fcgi_main
